requirements.txt
absl-py==2.1.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
easydict==1.11
facenet-pytorch==2.5.3
fonttools==4.47.2
google-auth==2.27.0
google-auth-oauthlib==1.0.0
graphviz==0.8.4
grpcio==1.60.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
joblib==1.3.2
kiwisolver==1.4.5
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.4
mxnet==1.9.1
mxnet-mkl==1.6.0
numpy==1.23.1
oauthlib==3.2.2
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
threadpoolctl==3.2.0
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchvision==0.13.1+cu113
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
Werkzeug==3.0.1
zipp==3.17.0
